# time: optimize Time.ISOWeek