--- a/src/time/time.go
+++ b/src/time/time.go
@@ -531,11 +531,18 @@ func absWeekday(abs uint64) Weekday {
 }
 
 // ISOWeek returns the ISO 8601 year and week number in which t occurs.
-// Week ranges from 1 to 53. Jan 01 to Jan 03 of year n might belong to
-// week 52 or 53 of year n-1, and Dec 29 to Dec 31 might belong to week 1
-// of year n+1.
+// according to the rule that the first calendar week of a calendar year
+// is the week including the first Thursday of that year, and that the last one
+// is the week immediately preceding the first calendar week of the next calendar year.
+// See https://www.iso.org/obp/ui#iso:std:iso:8601:-1:ed-1:v1:en:term:3.1.1.23 for details.
 func (t Time) ISOWeek() (year, week int) {
+	// begins on Monday and ends on Sunday
+	// Monday Tuesday Wednesday Thursday Friday Saturday Sunday
+	// 0      1       2         3        4      5        6
+	// +3     +2      +1        +0       -1     -2       -3
+	// the offset to Thursday
 	dd := 3 - (t.Weekday()+6)%7
+	// find the Thursday of calendar week
 	t = t.Add(Duration(dd) * 24 * Hour)
 	year, _, _, yday := t.date(false)
 	return year, yday/7 + 1