--- a/src/time/time_test.go
+++ b/src/time/time_test.go
@@ -437,24 +437,6 @@ func TestISOWeek(t *testing.T) {
 	}
 }
 
-func BenchmarkISOWeek(b *testing.B) {
-	// 7 days
-	tests := [...]Time{
-		Now().Add(1 * 24 * Hour),
-		Now().Add(2 * 24 * Hour),
-		Now().Add(3 * 24 * Hour),
-		Now().Add(4 * 24 * Hour),
-		Now().Add(5 * 24 * Hour),
-		Now().Add(6 * 24 * Hour),
-		Now().Add(7 * 24 * Hour),
-	}
-	for i := 0; i < b.N; i++ {
-		for _, tt := range tests {
-			tt.ISOWeek()
-		}
-	}
-}
-
 type YearDayTest struct {
 	year, month, day int
 	yday             int
@@ -1364,6 +1346,24 @@ func BenchmarkDay(b *testing.B) {
 	t := Now()
 	for i := 0; i < b.N; i++ {
 		_ = t.Day()
+	}
+}
+
+func BenchmarkISOWeek(b *testing.B) {
+	// 7 days
+	tests := [...]Time{
+		Now().Add(1 * 24 * Hour),
+		Now().Add(2 * 24 * Hour),
+		Now().Add(3 * 24 * Hour),
+		Now().Add(4 * 24 * Hour),
+		Now().Add(5 * 24 * Hour),
+		Now().Add(6 * 24 * Hour),
+		Now().Add(7 * 24 * Hour),
+	}
+	for i := 0; i < b.N; i++ {
+		for _, tt := range tests {
+			tt.ISOWeek()
+		}
 	}
 }