--- a/src/time/time_test.go
+++ b/src/time/time_test.go
@@ -1350,20 +1350,9 @@ func BenchmarkDay(b *testing.B) {
 }
 
 func BenchmarkISOWeek(b *testing.B) {
-	// 7 days
-	tests := [...]Time{
-		Now().Add(1 * 24 * Hour),
-		Now().Add(2 * 24 * Hour),
-		Now().Add(3 * 24 * Hour),
-		Now().Add(4 * 24 * Hour),
-		Now().Add(5 * 24 * Hour),
-		Now().Add(6 * 24 * Hour),
-		Now().Add(7 * 24 * Hour),
-	}
+	t := Now()
 	for i := 0; i < b.N; i++ {
-		for _, tt := range tests {
-			tt.ISOWeek()
-		}
+		_, _ = t.ISOWeek()
 	}
 }