@openset 2020-01-01
use abs more faster
```go
func (t Time) ISOWeek() (year, week int) {
	// weeks start with Monday
	// Monday Tuesday Wednesday Thursday Friday Saturday Sunday
	// 0      1       2         3        4      5        6
	// +3     +2      +1        0       -1     -2       -3
	// the offset to Thursday
	abs := t.abs()
	d := Thursday - absWeekday(abs)
	// handle Sunday
	if d == Thursday {
		d = -3
	}
	// find the Thursday of the calendar week
	d *= secondsPerDay
	if d > 0 {
		abs += uint64(d)
	} else if d < 0 {
		abs -= uint64(-d)
	}
	year, _, _, week = absDate(abs, false)
	return year, week/7 + 1
}
```